Add "secondary" channel mode args that are ramped with 4H.
This allows for a "saturation" parameter in HSV modes on 3-channel
lights. This uses 4H, thus removing momentary turbo from channels that
require a second channel arg.

This is a silly way to implement it; it would be nice to be more generic
and allow for some struct or arbitrary number of channel args or
something, but how many lights need more than 2 channel args, and you'd
need to do something like 23H to manage it all.